Refactor ListDetailsPage ListDetailsComponent to page-component ItemsListing
